requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
httpx>=0.28.0
google-genai>=1.0.0
pydantic-settings>=2.10.0
itsdangerous>=2.2.0
